invitations pt1.
- Add invitations table in backend
- Add api controller to create and list invites
- Add template concept of admin users in ui and api
- UI admin view (for invitiations) and menu item

TODO:
- conroler to delete invitaitons
- admin contorl of invitaitons
- regiration requiring invitiation code UI and API..

Now that there is a concept of admins (well 1, just me)
email verification should be looked at yet.
For now just keep most of the admin stuff pretty plain / not scary...
EG. deleting a wiki etc should not be in the UI, just via CLI? / with
some sort of confirmation?