mw db creation, switch to API code
This was it's own little JS service before.
But it is going to be eaiser to keep things going if this
is part of the same main codebase, even if the code / queue
etc is later run as a seperate process.

For now it runs with the main scheduler and in the main queue.