FIX API DB PROVISIONING
This was a stupid issue.

For some reason the CREATE USER always errors? even when the
user is actually created.
SO...
Run the create, but ignore errors, and then try to alter the password.
if the alter password fails, then the user was not created, if it
succeeds then the create was okay!