api, enable logo upload for a wiki
This currently writes to a wiki but doesnt write to the api db at all.

Resizes to a single logo which is appropriate for the top left of
mediawiki for now. Might want more (like a favicon) in the future).
We might also want that to be different?

Decided to just use a google cloud bucket for now, shouldnt be too hard
to move away from this to any other generic file store if and when we
want to...

This patch also adds a new service account for the api app to write to
the storage bucket using.

This is the start of #55