api, markdownify the account creation email
This used to be pure raw text email.
Now this actually has some style as we use a notification.

This also fixes an issue that might have been deployed with the upgrade
to 6.x laraval.
https://stackoverflow.com/questions/58162258/method-illuminate-translation-translatorgetfromjson-does-not-exist
This is probably because we pulled ResetPasswordNotification out of
laravel to modify it and it depends on Illuminate\Support\Facades\Lang.
MAybe I missed this in the release notes?