Add platform-nginx service to do fanout for apps
This removes the fanout logic from the ingress.
Each ingress resources was generating 700+ lines of configuration before
this.
Now it should only be a couple hundred..
Also in order to avoid the couple hundredyness, we might want to have a
single ingress for all custom domains, and instead just update the
domains that are in there....

This should give us more flexibility and the layering is now.

GCE LB -> K8s Ingress -> platform-nginx -> apps

For non app stuff it is still

GCE LB -> K8s Ingress -> service