135, make api make and use 135 dbs
This includes making the API only select dbs of a specific version
(rather than needing to use the unclaimed old ones)
And also makes sure the pool is populated for the new dbs too.

There are now 3 places that this db version is hardcoded in api code.
it should be an env var... and this whole thing should be in some
other service most likely....

For #118