Rework storage pool provisioning jobs
Instead of having 3 jobs this now just has 2.
Each job takes care of if the state of the pool, and if
we should need to create more.

This also means that at most 1 new entry will be created per
min. Which is fine for now, but might change in the future if
lots of things get created at once.