->fail() jobs rather than knowingly throwing exceptions
This was seen in https://github.com/wbstack/api/issues/11
when the api queue was running, but no query service was
yet up to create namespaces on.
The queue process would seemingly exit, as we threw an exception.
Looking at current docs we should be calling ->fail and passing
an exception.

This then allows the job to try to handle failures etc?

I added safe guard returns as the docs for this current laravel
version sucks a bit and it was unclear if execution really
stopped when calling fail?

Fixes #11