opportunistic storage pool repopulation
It's pretty easy to create a sandbox, so we could get
lots of them made in a short period of time.

In c259db095641660c27e5c8ef1a5c7a528e9b4a3e things
were re worked to mean these pools populate slower.
This should counteract the effects of that in a nice way

opportunistic jobs were always on the cards, even if they
should probably be done a little tidier than this...