Add observer to elasticsearch/lexeme settings
This adds an observer to the changes coming into all settings. It's
currently looking only for updated/created events on the lexeme or
elasticsearch setting and triggers the index update job if they are both
enabled at the same time. This change also makes the update job unique
to avoid duplicates.

I thought this could be a nice way to manage the lifecycle of these
settings in separate classes rather than stuffing everything into the
update controller.