[STANDALONE] Add DeleteWikiFinalizeJob (#158)
* Add DeleteWikiFinalizeJob

Adds a job that deletes the remainders of the wiki.

Delets GCS logos dir if necessary. Adds helper method on
WikiLogoController to get the dir.

To be run when everything else has been removed.

* Delete siteDirectory instead

- This moves the siteDir + logsDir static functions to Wiki
- Deletes the siteDirectory if exists rather than only the logosdir

* Check for elasticsearch index before deleting wiki

This introduces a ElasticSearchHelper to share the functionality of
querying ElasticSearch for any potential indices using the basename of
the wiki.