Adjustments to tests and functionality
The commit removes the usage of the download function
in place of an normal response call. So far we didn't
temoporarily save the csv to disk which is necessary for
using this method[1]. Instead this builds the csv in the
output buffer[2] and then sends the contents of that.

This therefore also adjusts the tests to remove the
dependency on the mocked download method and instead
relies on a single high-level test that not only checks
the status and the error but also loosely checks that some
of the expected content is there.

We also make some adjustments to parsing the data from
models which aren't guaranteed to be populated for all
wikis since the stats collection for them may not have run

Finally this commit slightly renames some of the headers
of the columns for clarity.

[1] https://laravel.com/docs/8.x/responses#file-downloads
[2] https://www.php.net/manual/en/book.outcontrol.php