Update to PHP 8.1 (#703)
* Update to PHP 8.1

As part of this process it was necssary to make some tweaks related
to our use of PDO calls. In 8.1 attempting to explicitly commit
or rollback a transaciton when none is in progress now errors
rather than failing silently[1].

We have a number of these in our wiki deletion code since the
altering of the tables automatically commits and ends the
transaction.

For this reason we are also unable use the traits that rollback
transactions after running tests if the tests themselves are
causing a commit. These have been removed too. Unfortunately
this adds more places where the tests are no longer idempotent.

[1] https://github.com/php/php-src/commit/990bb34891c83d12c5129fd781893704f948f2f4

* Run `composer update` after using php81 (#713)

* Run  after using php81

* Use legacy PSR-7 message factories

---------

Co-authored-by: Andrew Kostka <andrew.kostka@wikimedia.de>

---------

Co-authored-by: Andrew Kostka <andrew.kostka@wikimedia.de>