fix: Use MWTimestamps PlatformStatsSummaryJob
We appear to have overloooked that the DB format used
for MW revisions saved in MariaDB uses a special
string MWTimestampFormat.

This patch uses a previously introduced helper for
converting these to Carbon and then doing the necessary
comparisons.

It still uses the pattern of querying the MW dbs by
building a query from the api db which I believe in
the medium term we would like to move away from.

It adjust some comments to make this pattern more
understandable.

Bug: T364452