feat: rename active/inactive wiki metrics
There was some ambiguity about the meaning of
active and inactive so this commit more verbosely
renames them to exactly what they are: edited (or
not) in the last 90 days.

There is still a very mild unclarity in the name
not_edited_last_90_days because this excludes
empty wikis not edited in the last 90 days.

becuase not_edited_last_90_days_but_not_empty
seemed ludicrous.

Bug: T364452