Test create wiki api endpoint can accept a profile (#885)
This commit asserts that adding the additional profile
parameter to the wiki creation API does not result in
failure.

The purpose of this is to make clear that UI development
can go ahead and add this parameter to requests without
it being necessary to have completed the api work first.

It also adds assertions for the parameters that are required
for the wiki creation api

Bug: T388640