Introduce WikiProfile model and test
This commit introduces a basic wiki profile model, the migration
to save the data in the database and a very simple test to confirm
that it works.

I thought for a while about including a load of validation here
e.g. Other values have some string for the other details
but I decided that this should probably live just in a validator
that's used at the controller level.

I also decided that we would allow multiple profiles per wiki.
We'll need to just reference the newest one when showing them
to users but this will allow us to monitor the changes in profiles
over time.

Bug: T388895