Create WikiProfile objects when a Wiki is made (#905)
This creates a WikiProfile object if a profile is provided
at wiki creation time.

It also add some validation for the form of this profile
over the network.

It should be noted that the _other strings are stored
unsantised user input. We need to take this into account
if we do anything with them

Bug: T389063