Fix calls to deprecated version of PHPUnit providers (#990)
All PHPUnit Providers now need to be static and public[1].

This commit fixes the references to these in what I believe
was the simplest way possible. In ElasticSearchIndexInitTest
this meant it was necessary to move some mocking out of
the provider and into the test directly.

I investigated creating mocks statically; I discovered `mock()`
which creates a Mockery mock but I decided that making the
decision to move to a different mock library wasn't what I wanted
to do here. I just wanted my tests to run without deprecation
warnings.

[1] https://github.com/sebastianbergmann/phpunit/commit/9caafe2d49b33a21f87db248a8ad6ca7c7bdac09